requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
jinja2==3.1.2
openai==1.3.7
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0 